requirements.txt
feedparser
yagmail
newspaper3k
requests
lxml_html_clean